requirements.txt
feedparser>=6.0,<7
requests>=2.31,<3
python-dotenv>=1.0,<2
